# backend/config/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Average JS Webshop API",
    "description": "A backend API by Average JS Enjoyers",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/users/": {
      "get": {
        "tags": [
          "Admin"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Admin"
        ],
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  },
...
}